Better structure of the exercises actor and views
