+++ server/exercise/src/main/scala/com/eigengo/lift/exercise/ExerciseMarshallers.scala
@@ -1,11 +1,8 @@
 package com.eigengo.lift.exercise
 
-import java.util.UUID
-
 import com.eigengo.lift.common.{CommonMarshallers, CommonPathDirectives}
 import scodec.bits.BitVector
 import spray.http.HttpRequest
-import spray.httpx.marshalling.{ToResponseMarshallingContext, ToResponseMarshaller}
 import spray.httpx.unmarshalling.{Deserialized, FromRequestUnmarshaller}
 import spray.routing._
 import spray.routing.directives.{MarshallingDirectives, PathDirectives}
