+++ server/exercise/src/test/scala/com/eigengo/lift/exercise/ExerciseServiceTest.scala
@@ -5,9 +5,9 @@ import java.text.SimpleDateFormat
 import akka.actor.{ActorRef, ActorSystem}
 import akka.testkit.{ImplicitSender, TestActor, TestKitBase, TestProbe}
 import com.eigengo.lift.common.UserId
-import com.eigengo.lift.exercise.ExerciseClassifiers.{GetMuscleGroups, MuscleGroup}
-import com.eigengo.lift.exercise.UserExercises._
-import com.eigengo.lift.exercise.UserExercisesView._
+import com.eigengo.lift.exercise.UserExerciseClassifier.MuscleGroup
+import com.eigengo.lift.exercise.UserExercisesProcessor._
+import com.eigengo.lift.exercise.UserExercisesSessions._
 import org.scalatest.{FlatSpec, Matchers}
 import scodec.bits.BitVector
 import spray.testkit.ScalatestRouteTest
@@ -41,9 +41,6 @@ object ExerciseServiceTest {
     probe.setAutoPilot {
       new TestActor.AutoPilot {
         def run(sender: ActorRef, msg: Any) = msg match {
-          case GetMuscleGroups =>
-            sender ! TestData.muscleGroups
-            TestActor.KeepRunning
           case UserExerciseSessionStart(_, _) =>
             sender ! \/.right(TestData.userId.id)
             TestActor.KeepRunning
@@ -90,16 +87,14 @@ class ExerciseServiceTest
 
   val probe = ExerciseServiceTest.probe
 
-  val underTest = exerciseRoute(probe.ref, probe.ref, probe.ref)
+  val underTest = exerciseRoute(probe.ref, probe.ref)
 
   val dateFormat = new SimpleDateFormat("yyyy-MM-dd")
 
   "The Exercise service" should "listen at GET /exercise/musclegroups endpoint" in {
     Get("/exercise/musclegroups") ~> underTest ~> check {
-      responseAs[List[MuscleGroup]] should be(TestData.muscleGroups)
+      responseAs[List[MuscleGroup]] should be(UserExerciseClassifier.supportedMuscleGroups)
     }
-
-    probe.expectMsg(GetMuscleGroups)
   }
 
   it should "listen at POST exercise/:UserIdValue endpoint" in {
